Refactoring
-Realized having a package for each challenge was dumb
-Created JUnit testcase
-Refactored the first class to not do its task upon instantiation,
prints out and returns the answer upon calling method (another dumb
idea)
-Some other stuff probably, just housekeeping
