+++ src/projectEuler/ProjectEuler002.java
@@ -1,0 +1,5 @@
+package projectEuler;
+
+public class ProjectEuler002 {
+	
+}
