+++ src/test/Test.java
@@ -1,10 +1,12 @@
 package test;
 
-import pe001.ProjectEuler001;
+import projectEuler.ProjectEuler001;
 
 /**
  * 
- * Contains main method to test projects. Each project should perform its task automatically on instantiation
+ * Contains main method to test challenges. Each class should perform its task automatically on instantiation.
+ * 
+ * 
  * 
  * @author Alex
  * 
@@ -12,6 +14,6 @@ import pe001.ProjectEuler001;
 public class Test {
 	public static void main(String[] args){
 		ProjectEuler001 pe001 = new ProjectEuler001();
-		System.out.println(pe001.findSumOfMultiples(10));
+		pe001.findSumOfMultiples();
 	}
 }
